+++ app/src/main/java/ru/uniquenature/myapplication/data/Movie.kt
@@ -5,26 +5,26 @@ import android.os.Parcelable
 
 @Suppress("UNREACHABLE_CODE")
 data class Movie(
-        val id: Int,
+        val id: Long,
         val title: String?,
         val overview: String?,
         val poster: String?,
         val backdrop: String?,
-        val ratings: Float,
-        val numberOfRatings: Int,
+        val ratings: Double,
+        val numberOfRatings: Long,
         val minimumAge: Int,
         val runtime: Int,
         val genres: List<Genre>,
         val actors: List<Actor>
 ):Parcelable {
     constructor(parcel: Parcel) : this(
-            parcel.readInt(),
+            parcel.readInt().toLong(),
             parcel.readString(),
             parcel.readString(),
             parcel.readString(),
             parcel.readString(),
-            parcel.readFloat(),
-            parcel.readInt(),
+            parcel.readFloat().toDouble(),
+            parcel.readInt().toLong(),
             parcel.readInt(),
             parcel.readInt(),
             genres = TODO("genres"),
@@ -32,13 +32,13 @@ data class Movie(
     )
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeInt(id)
+        parcel.writeInt(id.toInt())
         parcel.writeString(title)
         parcel.writeString(overview)
         parcel.writeString(poster)
         parcel.writeString(backdrop)
-        parcel.writeFloat(ratings)
-        parcel.writeInt(numberOfRatings)
+        parcel.writeFloat(ratings.toFloat())
+        parcel.writeInt(numberOfRatings.toInt())
         parcel.writeInt(minimumAge)
         parcel.writeInt(runtime)
     }
